--- a/core/dnsserver/address_test.go
+++ b/core/dnsserver/address_test.go
@@ -45,8 +45,9 @@ func TestNormalizeZoneReverse(t *testing.T) {
 		{"10.0.0.0/24.:53", "dns://10.0.0.0/24.:53", false},
 
 		// non %8==0 netmasks
-		{"2003::53/67", "dns://2003::53/67.:53", false},
-		{"10.0.0.0/25.", "dns://10.0.0.0/25.:53", false},
+		{"2003::53/67", "dns://0.0.0.0.0.0.0.0.0.0.0.0.0.0.0.0.0.0.0.0.3.0.0.2.ip6.arpa.:53", false},
+		{"10.0.0.0/25.", "dns://10.0.0.0/25.:53", false}, // has dot
+		{"10.0.0.0/25", "dns://0.0.10.in-addr.arpa.:53", false},
 	} {
 		addr, err := normalizeZone(test.input)
 		actual := addr.String()