### 1. What does this pull request do?
Enables CIDR routing of queries on non `x%8==0` boundaries.

### 2. Which issues (if any) are related?
Fixes #1074

### 3. Which documentation changes (if any) need to be made?
Yes, this needs to be documented, but it mostly should just work as intended.  But all docs on coredns.io need updating to this.