@miekg 2017-10-23
Probably can do w/ some extra corner case testing, but main code should be good to go.

(/cc @chrisohaver / @johnbelamaric )