@chrisohaver 2017-10-23
Looks good.  One thing, perhaps filter func should take query string and query type, so the func can return before trying to extract addresses from non PTR request query string...