@miekg 2017-10-23
Good point and I thought about that as well. But a reverse zone is just a normal DNS zone and SOA and NS queries are valid as well (or whatever you put in there).

Making FilterFunc more general is thing we can do of course, but I like doing the minimal thing for the current problem.