@chrisohaver 2017-10-23
Ah, I wasn't trying to generalize the function. I was trying to suggest an optimization.  In the proposed implementation, every query, regardless of type will get run through `dnsutil.ExtractAddressFromReverse`...

I have only seen PTR types for reverse queries.  Can you give an example of a non-PTR reverse zone query?