@miekg 2017-10-23
~~~
 dig -x 8.8.8.8 soa                                                                                                              ~/g/src/github.com/coredns/coredns/test classless

; <<>> DiG 9.10.3-P4-Debian <<>> -x 8.8.8.8 soa
;; global options: +cmd
;; Got answer:
;; ->>HEADER<<- opcode: QUERY, status: NOERROR, id: 47247
;; flags: qr rd ra; QUERY: 1, ANSWER: 0, AUTHORITY: 1, ADDITIONAL: 0

;; QUESTION SECTION:
;8.8.8.8.in-addr.arpa.		IN	SOA

;; AUTHORITY SECTION:
8.8.8.in-addr.arpa.	60	IN	SOA	ns2.google.com. dns-admin.google.com. 173100588 900 900 1800 60
~~~