@miekg 2017-10-24
ExtractReverseFrom is broken because it checks  the validity of the address, which it shouldn't