@miekg 2017-10-24
grr. dns.ReverseAddr does the same.